llm_service/requirements.txt
llama-cpp-python
paho-mqtt>=1.6.1
python-dotenv>=1.0.1
